misc fixes
zig will still spit out a error with "TODO"
Seems to be triggered by creating a scheduler type that has events
with a function directly instead of registered through a struct.

Since this is needed for DependOn among other things this error will
persist until the zig compiler finishes whatever "TODO" is triggering
this issue. Alternative would be to resolve issue #70 "DependOn should
accept system structs as well". This could then be used as a temporary
workaround.