scheduler: update to 0.15, nuke thread pool ...
This is a substantial commit. With these changes ecez runs with zig
0.15. The patch also removes the custom thread pool and uses the std
version instead. The current code is designed to fit with the future
async/await model that is expected in 0.16 and we can drop in replace
the use of the thread pool with the future async!

To achieve this the API is changed. You now are expected to declare a
scheduler before initializing it. You will also need the storage type to
create the scheduler type which is probably the biggest painpoint.
Finally, you will also need any EventArgument types to create the
scheduler.