storage: implement createEntity with setComponents
Instead of creating entity with ensureCapacity +
createEntityAssumeCapacity which boils down to atomic load + fetchAdd,
use a single fetchAdd (or fetch an inactive entity), then use
setComponents to populate the entity.

createEntity will also handle OOM properly and if all fails, report the
total failure to the application (which may choose to panic or change
allocation strategy if needed).

This resolved #293 'storage: ensureUnusedCapacity +
createEntityAssumeCapacity (= createEntity) is racy?'

Co-authored-by: Simon Galli <hubi@hubidubi.net>