storage: CreateStorage slice instead of tuple
Use a slice of types instead of a tuple. This simplifies the process of
concatinating components for the final CreateStorage input. It also
simplifies a few things internally.

This commit is breaking.

Resolves #151 'GroupComponents function: Allow registering multiple
components in a single tuple/array'

Resolved #294 'CreateStorage should get a slice of component types
instead of tuple'